Move fixedvalue to base.utils
